requirement.txt
Flask==3.0.3
flask-cors==6.0.2
google-generativeai==0.8.3
python-dotenv==1.0.1